Add function “bitboard_stat_to_code”
The function returns code for side singles and mams.
The code is bigger if total checkers amount is bigger.
The code is bigger if checkers amount is equal and mam count is bigger.
No argument checks.

Here is an example for some values.
King counts are in rows.
Single count are in columns.

     0  1  2  3  4  6
--+-------------------
0 |  -  1  4  8 13 19
1 |  0  3  7 12 18
2 |  2  6 11 17
3 |  5 10 16
4 |  9 15
5 | 14