Implement RusCheckers.check logic.
It is quite complicate. Sometimes move may be absent in
“move list” output command. And “new_fen” may differs from
“fen” output.

So, our first step is normalizing a fen (set/get fen).
“new_fen” will be built in rus-checkers.

Second step is bulding move lists. It contains one item if
move is found in move list. It contains all items if not found.

Third step is attemption to try every move in moves list.
We return True if new_fen is the same as passed.