Add C1 and C2 range instead C for dynamic changing.
The main idea that it would be optimal to change C during simulations.
At the beginning of a simulation it would be like to have large C to
increase variability. At the and of a simulation it would be better
to have smaller C, because selecting moves with small qsimulations cannot
affect best move changing.