Tweak move printing format in MCTS explanation.
Old version printed only the first and the last squares of a move.
It is understandable for humon in most cases. But another command
prints full move with all inner cells during taking. So annoying
errors is possible if explanation is used in another applications.
So now explanation worked a little bit slowly but it prints all
inner squares during taking.